api-added: interface to /update to get data
added a new table device and changed the relational shape of the data.
Now Device should be registered initially through /registerdevice with a
serial number. Then user can be associated with the device and each
request from the user is stored in Location with the device id.
/update exposes an uri through which we are able to request the pi to
fetch any incoming data. On receiving an incoming data, js forwards with
the location data to a new page.